Rewrote convert.php
1. The original script only takes a copy of the decoded json file and
decode it again (which makes no sense). The new script parses a large
json file and retrives the id, url, user and caption of each image in
the json file. It then uses the retrived data to create an image object
for each image in the json file. The convert() function should be called
by cache with the json file to be parsed.  This function will then
return an array of image objects. 2. Added comments in each file to
indicate the author.